Move the services into their own file like we do in other products
